Address PR review feedback: drop SafeObserver, scope try/catch to OnFinalized
Eric's review surfaced two related issues:

- TelemetryObserver's internal Safe() helper and SafeObserver were
  byte-identical except for the source tag; "neither wraps the other"
  was silent duplication that would drift over time.
- Wrapping plain field-assignment lifecycle methods in try/catch was
  paranoia. The only code in this class that can realistically throw is
  the proto build + telemetry-client enqueue inside OnFinalized.

Resolution:

- Delete SafeObserver.cs and SafeObserverTests.cs. The "optional for
  future third-party observers" framing was the weaker of the two paths
  Eric offered; third-party observers self-guard like the first-party
  ones do.
- Remove Safe(Action) from TelemetryObserver. Lifecycle methods are now
  straight scalar writes with an IsFinalized() short-circuit at the top.
- Scope the try/catch to the genuine risk surface inside OnFinalized
  (BuildTelemetryLog + client.Enqueue), emitting the same suppression
  activity event on failure.
- SafeMessage stays because OnError still has to neutralize the
  exception-with-throwing-.Message case before storing it on the context.
- Update the design doc and sprint plan so they no longer reference the
  removed class.
